+++ src/lib/audio/audioCorrelation.h
@@ -60,7 +60,7 @@ public:
       Correlates the two vectors envMain and envSub.
       \c correlation must be a pre-allocated vector of size sizeMain+sizeSub+1.
       */
-    static void correlate(const qint64 *envMain, int sizeMain, const qint64 *envSub, int sizeSub, qint64 *correlation, qint64 *out_max = nullptr);
+    static void correlate(const qint64 *envMain, size_t sizeMain, const qint64 *envSub, size_t sizeSub, qint64 *correlation, qint64 *out_max = nullptr);
 
 private:
     std::unique_ptr<AudioEnvelope> m_mainTrackEnvelope;
@@ -69,13 +69,13 @@ private:
     QList<AudioCorrelationInfo *> m_correlations;
 
 private slots:
-     /**
-      This is invoked when the child envelope is computed. This
-      triggers the actual computations of the cross-correlation for
-      aligning the envelope to the reference envelope.
+    /**
+     This is invoked when the child envelope is computed. This
+     triggers the actual computations of the cross-correlation for
+     aligning the envelope to the reference envelope.
 
-      Takes ownership of @p envelope.
-    */
+     Takes ownership of @p envelope.
+   */
     void slotProcessChild(AudioEnvelope *envelope);
     void slotAnnounceEnvelope();
